Abort xhr on cancel
Exposes jquery.xhr.abort on the deferred object.

Also clearing the hasIdentifiers cache, but the crux of the issue was `fetch` on the collection prototype was not cancelable, so preventing the `search` trigger only cancelled some things, but the collection continued to update.

Also clearing the debounce for good measure